requirements.txt
django==1.10
dj-database-url==0.4.1
django-import-export==0.4.5
pytz==2016.6.1
jinja2==2.8
psycopg2==2.6.2
django-webpack-loader==0.3.3
whitenoise==3.2.1
python-cjson==1.1.0

parsel==1.0.3
selenium==2.53.6
requests==2.11.0
python-dateutil==2.5.3
lxml==3.6.1

ansible==2.1.1
honcho==0.7.1
gunicorn==19.6.0
pymongo==3.3.0

flake8==3.0.4
ipython==5.0.0
ipdb==0.10.1
django-debug-toolbar==1.5
